# 93/PR 3/3 - Propagate performance optimisation throughout checks & validation functions