--- a/DESCRIPTION
+++ b/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: hubValidations
 Title: Testing framework for hubverse hub validations
-Version: 0.4.0
+Version: 0.5.0
 Authors@R: c(
     person(
         given = "Anna",