--- a/R/check_tbl_spl_non_compound_tid.R
+++ b/R/check_tbl_spl_non_compound_tid.R
@@ -17,7 +17,8 @@
 #' for more details.
 #' @export
 check_tbl_spl_non_compound_tid <- function(tbl, round_id, file_path, hub_path,
-                                           compound_taskid_set = NULL) {
+                                           compound_taskid_set = NULL,
+                                           derived_task_ids = NULL) {
   if (!is.null(compound_taskid_set) && isTRUE(is.na(compound_taskid_set))) {
     cli::cli_abort("Valid {.var compound_taskid_set} must be provided.")
   }
@@ -33,7 +34,8 @@ check_tbl_spl_non_compound_tid <- function(tbl, round_id, file_path, hub_path,
     return(skip_v3_spl_check(file_path))
   }
 
-  hash_tbl <- spl_hash_tbl(tbl, round_id, config_tasks, compound_taskid_set)
+  hash_tbl <- spl_hash_tbl(tbl, round_id, config_tasks, compound_taskid_set,
+                           derived_task_ids = derived_task_ids)
 
   n_tbl <- dplyr::summarise(
     hash_tbl,