… NAMESPACE
… NEWS.md
… R/match_tbl_to_model_task.R
… R/submission_tmpl.R
… R/v3-sample-utils.R
… R/validate_pr.R
… R/validate_submission.R
… _pkgdown.yml
… man/check_tbl_spl_compound_taskid_set.Rd
… man/check_tbl_spl_compound_tid.Rd
… man/check_tbl_spl_n.Rd
… man/check_tbl_spl_non_compound_tid.Rd
… man/check_tbl_value_col.Rd
… man/check_tbl_values.Rd
… man/check_tbl_values_required.Rd
… man/get_tbl_compound_taskid_set.Rd
… man/match_tbl_to_model_task.Rd
… man/submission_tmpl.Rd
… man/validate_model_data.Rd
… man/validate_pr.Rd
… man/validate_submission.Rd
… tests/testthat/_snaps/check_tbl_spl_compound_taskid_set.md
… tests/testthat/_snaps/check_tbl_spl_compound_tid.md
… tests/testthat/_snaps/check_tbl_spl_n.md
… tests/testthat/_snaps/check_tbl_spl_non_compound_tid.md
… tests/testthat/_snaps/check_tbl_value_col.md
… tests/testthat/_snaps/check_tbl_values.md
… tests/testthat/_snaps/check_tbl_values_required.md
… tests/testthat/_snaps/expand_model_out_grid.md
… tests/testthat/_snaps/match_tbl_to_model_task.md